+++ deploy/view_carbon_tax_calculation.sql
@@ -7,7 +7,7 @@ create or replace view ggircs.carbon_tax_calculation as
     with x as (
         select fuel.fuel_type                           as fuel_type,
                fuel.annual_fuel_amount                  as amount,
-               report.reporting_period_duration         as rpd,
+               report.reporting_period_duration::integer         as rpd,
                pro_rated_carbon_tax_rate                as pro_rated_ctr,
                pro_rated_implied_emission_factor        as pro_rated_ief
         from ggircs_swrs.fuel
@@ -17,10 +17,10 @@ create or replace view ggircs.carbon_tax_calculation as
                     on fuel.ghgr_import_id = report.ghgr_import_id
                 join ggircs.pro_rated_carbon_tax_rate as ctr
                     on fuel.fuel_type = ctr.fuel_type
-                    and report.reporting_period_duration = ctr.reporting_year
+                    and report.reporting_period_duration::integer = ctr.reporting_year
                 join ggircs.pro_rated_implied_emission_factor as ief
                     on ief.fuel_mapping_id = fm.id
-                    and report.reporting_period_duration = ief.reporting_year
+                    and report.reporting_period_duration::integer = ief.reporting_year
 
     )
     select x.rpd as year,
