+++ test/unit/view_pro_rated_carbon_tax_rate_test.sql
@@ -150,7 +150,7 @@ select results_eq(
     'select reporting_year from ggircs.pro_rated_carbon_tax_rate order by reporting_year',
 
     $$
-    select reporting_period_duration
+    select reporting_period_duration::integer
     from ggircs_swrs.fuel as fuel
     join ggircs_swrs.report as report
     on report.ghgr_import_id = fuel.ghgr_import_id
@@ -170,7 +170,7 @@ select results_eq(
     join ggircs_swrs.report as report
     on report.ghgr_import_id = fuel.ghgr_import_id
     )
-    select concat((x.rpd)::text, '-12-31')::date - concat((x.rpd)::text, '-01-01')::date as year_length
+    select concat((x.rpd::integer)::text, '-12-31')::date - concat((x.rpd::integer)::text, '-01-01')::date as year_length
     from x where x.fuel_type = 'Wood Waste'
     $$,
 
@@ -187,13 +187,13 @@ select results_eq(
     join ggircs_swrs.report as report
         on report.ghgr_import_id = fuel.ghgr_import_id
     join ggircs_swrs.carbon_tax_rate_mapping as ctr
-        on concat((report.reporting_period_duration - 1)::text, '-04-01')::date >= ctr.rate_start_date
+        on concat((report.reporting_period_duration::integer - 1)::text, '-04-01')::date >= ctr.rate_start_date
         and concat(report.reporting_period_duration::text, '-03-31')::date <= ctr.rate_end_date
     )
     select
            case
-               when x.rpd <= 2017 then 0
-               when x.rpd > 2021 then (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = (select max(id) from ggircs_swrs.carbon_tax_rate_mapping))
+               when x.rpd::integer <= 2017 then 0
+               when x.rpd::integer > 2021 then (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = (select max(id) from ggircs_swrs.carbon_tax_rate_mapping))
                else x.rate
            end as start_rate
     from x where x.fuel_type = 'Wood Waste' order by start_rate desc
@@ -212,17 +212,17 @@ select results_eq(
     join ggircs_swrs.report as report
         on report.ghgr_import_id = fuel.ghgr_import_id
     join ggircs_swrs.carbon_tax_rate_mapping as ctr
-        on concat((report.reporting_period_duration - 1)::text, '-04-01')::date >= ctr.rate_start_date
+        on concat((report.reporting_period_duration::integer - 1)::text, '-04-01')::date >= ctr.rate_start_date
         and concat(report.reporting_period_duration::text, '-03-31')::date <= ctr.rate_end_date
     )
 
     select
         case
-            when x.rpd <= 2017 then 0
-            when x.rpd > 2021 then concat((x.rpd)::text, '-04-01')::date - concat((x.rpd)::text, '-01-01')::date
+            when x.rpd::integer <= 2017 then 0
+            when x.rpd::integer > 2021 then concat((x.rpd::integer)::text, '-04-01')::date - concat((x.rpd::integer)::text, '-01-01')::date
             else (select rate_start_date
                   from ggircs_swrs.carbon_tax_rate_mapping
-                  where id = x.id+1) - concat((x.rpd)::text, '-01-01')::date
+                  where id = x.id+1) - concat((x.rpd::integer)::text, '-01-01')::date
         end as start_duration
     from x where x.fuel_type = 'Wood Waste' order by start_duration asc
     $$,
@@ -240,14 +240,14 @@ select results_eq(
     join ggircs_swrs.report as report
         on report.ghgr_import_id = fuel.ghgr_import_id
     join ggircs_swrs.carbon_tax_rate_mapping as ctr
-        on concat((report.reporting_period_duration - 1)::text, '-04-01')::date >= ctr.rate_start_date
+        on concat((report.reporting_period_duration::integer - 1)::text, '-04-01')::date >= ctr.rate_start_date
         and concat(report.reporting_period_duration::text, '-03-31')::date <= ctr.rate_end_date
     )
     select
            case
-               when x.rpd < 2017 then 0
-               when x.rpd = 2017 then 30
-               when x.rpd > 2021 then (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = (select max(id) from ggircs_swrs.carbon_tax_rate_mapping))
+               when x.rpd::integer < 2017 then 0
+               when x.rpd::integer = 2017 then 30
+               when x.rpd::integer > 2021 then (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = (select max(id) from ggircs_swrs.carbon_tax_rate_mapping))
                else (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = x.id+1)
            end as end_rate
     from x where x.fuel_type = 'Wood Waste' order by end_rate desc
@@ -266,15 +266,15 @@ select results_eq(
     join ggircs_swrs.report as report
         on report.ghgr_import_id = fuel.ghgr_import_id
     join ggircs_swrs.carbon_tax_rate_mapping as ctr
-        on concat((report.reporting_period_duration - 1)::text, '-04-01')::date >= ctr.rate_start_date
+        on concat((report.reporting_period_duration::integer - 1)::text, '-04-01')::date >= ctr.rate_start_date
         and concat(report.reporting_period_duration::text, '-03-31')::date <= ctr.rate_end_date
     )
 
     select case
-               when x.rpd < 2017 then 0
-               when x.rpd = 2017 then '2017-12-31'::date - '2017-04-01'::date
-               when x.rpd > 2021 then concat((x.rpd)::text, '-12-31')::date - concat((x.rpd)::text, '-04-01')::date
-               else concat((x.rpd)::text, '-12-31')::date - (select rate_start_date
+               when x.rpd::integer < 2017 then 0
+               when x.rpd::integer = 2017 then '2017-12-31'::date - '2017-04-01'::date
+               when x.rpd::integer > 2021 then concat((x.rpd::integer)::text, '-12-31')::date - concat((x.rpd::integer)::text, '-04-01')::date
+               else concat((x.rpd::integer)::text, '-12-31')::date - (select rate_start_date
                                                              from ggircs_swrs.carbon_tax_rate_mapping
                                                              where id = x.id+1)
            end as end_duration
@@ -299,42 +299,42 @@ select results_eq(
                  join ggircs_swrs.report as report
                       on fuel.ghgr_import_id = report.ghgr_import_id
                  join ggircs_swrs.carbon_tax_rate_mapping as ctr
-                      on concat((report.reporting_period_duration - 1)::text, '-04-01')::date >= ctr.rate_start_date
+                      on concat((report.reporting_period_duration::integer - 1)::text, '-04-01')::date >= ctr.rate_start_date
                       and concat(report.reporting_period_duration::text, '-03-31')::date <= ctr.rate_end_date
     ), y as (
-    select x.rpd as reporting_year,
+    select x.rpd::integer as reporting_year,
            x.fuel_type as fuel_type,
            case
-               when x.rpd <= 2017 then 0
-               when x.rpd > 2021 then (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = (select max(id) from ggircs_swrs.carbon_tax_rate_mapping))
+               when x.rpd::integer <= 2017 then 0
+               when x.rpd::integer > 2021 then (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = (select max(id) from ggircs_swrs.carbon_tax_rate_mapping))
                else x.rate
            end as start_rate,
 
            case
-               when x.rpd < 2017 then 0
-               when x.rpd = 2017 then 30
-               when x.rpd > 2021 then (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = (select max(id) from ggircs_swrs.carbon_tax_rate_mapping))
+               when x.rpd::integer < 2017 then 0
+               when x.rpd::integer = 2017 then 30
+               when x.rpd::integer > 2021 then (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = (select max(id) from ggircs_swrs.carbon_tax_rate_mapping))
                else (select carbon_tax_rate from ggircs_swrs.carbon_tax_rate_mapping where id = x.id+1)
            end as end_rate,
 
            case
-               when x.rpd <= 2017 then 0
-               when x.rpd > 2021 then concat((x.rpd)::text, '-04-01')::date - concat((x.rpd)::text, '-01-01')::date
+               when x.rpd::integer <= 2017 then 0
+               when x.rpd::integer > 2021 then concat((x.rpd::integer)::text, '-04-01')::date - concat((x.rpd::integer)::text, '-01-01')::date
                else (select rate_start_date
                      from ggircs_swrs.carbon_tax_rate_mapping
-                     where id = x.id+1) - concat((x.rpd)::text, '-01-01')::date
+                     where id = x.id+1) - concat((x.rpd::integer)::text, '-01-01')::date
            end as start_duration,
 
            case
-               when x.rpd < 2017 then 0
-               when x.rpd = 2017 then '2017-12-31'::date - '2017-04-01'::date
-               when x.rpd > 2021 then concat((x.rpd)::text, '-12-31')::date - concat((x.rpd)::text, '-04-01')::date
-               else concat((x.rpd)::text, '-12-31')::date - (select rate_start_date
+               when x.rpd::integer < 2017 then 0
+               when x.rpd::integer = 2017 then '2017-12-31'::date - '2017-04-01'::date
+               when x.rpd::integer > 2021 then concat((x.rpd::integer)::text, '-12-31')::date - concat((x.rpd::integer)::text, '-04-01')::date
+               else concat((x.rpd::integer)::text, '-12-31')::date - (select rate_start_date
                                                              from ggircs_swrs.carbon_tax_rate_mapping
                                                              where id = x.id+1)
            end as end_duration,
 
-           concat((x.rpd)::text, '-12-31')::date - concat((x.rpd)::text, '-01-01')::date as year_length
+           concat((x.rpd::integer)::text, '-12-31')::date - concat((x.rpd::integer)::text, '-01-01')::date as year_length
     from x)
     select
            ((y.start_rate * y.start_duration) + (y.end_rate * y.end_duration)) / y.year_length as pro_rated_carbon_tax_rate
